Abril 07 3

+++ app/Http/Controllers/Modulos/Sitios/SitiosController.php
@@ -20,7 +20,7 @@ class SitiosController extends controller {
         $nombre = "Administrador";
         $categorias = DB::select("SELECT * FROM bdp_categoria");
         $sub_subcategorias = array();
-        $sitio = DB::select("SELECT * FROM bdp_sitio WHERE sit_id = ?", array($id));
+        $sitio = DB::select("SELECT * FROM bdp_sitio WHERE sit_deleted_at = NULL AND est_id = '1' AND sit_id = ?", array($id));
         $img = DB::select("SELECT img_ruta FROM bdp_imagen WHERE sit_id = ?", array($id));
 
         $sitio = $sitio[0];
@@ -41,23 +41,32 @@ class SitiosController extends controller {
     }
 
     function getIndex() {
-        $sitios = DB::select("SELECT * FROM bdp_sitio, bdp_imagen WHERE bdp_imagen.sit_id = bdp_sitio.sit_id");
+        
+        $sitios = DB::select("SELECT * FROM bdp_sitio, bdp_imagen WHERE sit_deleted_at IS NULL AND est_id = '1' AND bdp_imagen.sit_id = bdp_sitio.sit_id");
         //$sitios = $sitios[0];
 
         return view('Modulos.Sitios.sitio', compact('sitios'));
     }
 
     function getIndexxx() {
-
+ if (Session::has('user') === true){
+        
         $nombre = "Administrador";
         $categorias = DB::select("SELECT * FROM bdp_categoria");
         $usuarios = DB::select("SELECT usu_id, usu_usuario FROM bdp_usuario WHERE usu_deleted_at IS NULL AND usu_activado = '1' ");
         $sub_subcategorias = array();
         return view('Modulos.Sitios.crearsitio', compact("categorias", "subcategorias", "sub_subcategorias", "usuarios"));
+         
+         }else{
+            echo 'Permiso denegado';
+            return redirect(url("index"));
+        }
     }
 
     function postIndexxx() {
-
+ if (Session::has('user') === true){
+     
+      
         $sitio = filter_input_array(INPUT_POST)['sitio'];
         extract($sitio);
 
@@ -141,26 +150,43 @@ class SitiosController extends controller {
 
 
 
-        Session::put('success', 'Usuario creado exitosamente');
+       // Session::put('success', 'Usuario creado exitosamente');
 
 //    DB::insert("INSERT INTO usuario (nombre, apellido) VALUES(?,?)", array($nombre, $apellido));
-            return redirect(url("/sitios/sitios/index") );
+        
+        
+        return redirect(url("/sitios/sitios/listar") );
+             
+             
+         }else{
+            echo 'Permiso denegado';
+            return redirect(url("index"));
+        }
     }
 
     function getListar() {
+         if (Session::has('user') === true){
         
-         if(Session::get('rol') == 3){
+         if (Session::has('super') === true){
+         
         $sitios = DB::select("SELECT * FROM bdp_sitio");
          }else{
-             $sitios = DB::select("SELECT * FROM bdp_sitio WHERE sit_deleted_at IS NULL AND usu_id = ?", array(  Session::get('rol')  ) );
+             $sitios = DB::select("SELECT * FROM bdp_sitio WHERE sit_deleted_at IS NULL AND usu_id = ?", array(  Session::get('user')  ) );
              //$usuarioI = DB::select("SELECT usu_id FROM bdp_usuario WHERE usu_deleted_at IS NULL AND usu_activado = '1' AND (usu_usuario = ? AND usu_password = ?)", array($user, $pass));
             
          }
          
         return view('Modulos.Sitios.listarsitio', compact("sitios"));
+         
+         }else{
+            echo 'Permiso denegado';
+            return redirect(url("index"));
+        }
     }
 
     function getUpdate($id) {
+         if (Session::has('user') === true){
+      
 //        $sitio = filter_input_array(INPUT_POST)['sitio'];
 //        extract($sitio);
 
@@ -170,9 +196,17 @@ class SitiosController extends controller {
         $sitio = DB::select("SELECT * FROM bdp_sitio WHERE sit_id = ?", array($id));
         $sitio = $sitio[0];
         return view('Modulos.Sitios.modificarsitio', compact("categorias", "subcategorias", "sub_subcategorias", "sitio"));
+         
+        }else{
+            echo 'Permiso denegado';
+            return redirect(url("index"));
+        }
     }
 
     function postUpdate() {
+         if (Session::has('user') === true){
+       
+    
         $target_dir = dirname(getcwd());
 
         $target_file = $target_dir . basename($_FILES["fileToUpload"]["name"]);
@@ -224,9 +258,16 @@ class SitiosController extends controller {
 
         Session::put('success', 'Sitios modificado exitosamente');
         return redirect(url("/sitios/sitios/listar"));
+         
+        }else{
+            echo 'Permiso denegado';
+            return redirect(url("index"));
+        }
     }
 
     function getSuspender($id) {
+         if (Session::has('user') === true){
+       
         $sit_deleted_at = date("Y-m-d\TH:i:s");
 
 
@@ -234,14 +275,45 @@ class SitiosController extends controller {
                 . " WHERE sit_id = ?", array($sit_deleted_at, $id));
 
         return redirect(url("sitios/sitios/listar"));
+         }else{
+            echo 'Permiso denegado';
+            return redirect(url("index"));
+        }
     }
 
     function getEliminar($id) {
+         if (Session::has('user') === true){
+             
+         
+//        DB::delete("DELETE FROM bdp_imagen WHERE sit_id = ?", array($id));
+//        DB::delete("DELETE FROM bdp_sitio WHERE sit_id = ?", array($id));
+        // DB::update("UPDATE bdp_dato_usuario SET dus_deleted_at = ? WHERE usu_id = ?", array( $usu_deleted_at, $id));
+
+        //return redirect(url("sitios/sitios/listar"));
+        return view('Modulos.Sitios.eliminar', compact("id", "dato_usuario"));
+         
+        }else{
+            echo 'Permiso denegado';
+            return redirect(url("index"));
+        }
+    }
+    
+    function getBorrar($id) {
+       
+     
+        if (Session::has('super') === true) {
+             
+         
         DB::delete("DELETE FROM bdp_imagen WHERE sit_id = ?", array($id));
         DB::delete("DELETE FROM bdp_sitio WHERE sit_id = ?", array($id));
         // DB::update("UPDATE bdp_dato_usuario SET dus_deleted_at = ? WHERE usu_id = ?", array( $usu_deleted_at, $id));
 
         return redirect(url("sitios/sitios/listar"));
+        
+        }else{
+            echo 'Permiso denegado';
+            return redirect(url("index"));
+        }
     }
 
 }
